reservation-service/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
drf-yasg==1.20.0
requests==2.25.1
watchdog==2.1.2
django-cors-headers==3.7.0
pydantic==1.8.2